Add files via upload
This is a overarching program I made that opens up the other programs I've made to help automate work.  The other programs are called in as import statements within this code and activate when you select them in the GUI.  This is purely on here to show an example of the number of programs I've made to help automate work.